+++ e2e/common/utilities/component.util.service.ts
@@ -1,11 +1,11 @@
-import { by, ElementFinder, promise } from 'protractor';
+import { by, ElementFinder } from 'protractor';
 
-// TODO: Move components.util.service to the systelab-components. Split this file into utils for each component
+// TODO: Split this file into utils for each component
 export class ComponentUtilService {
-	public static readonly tagNameContextualMenu = 'systelab-context-menu-item';
-	public static readonly classNameDropDown = 'slab-dropdown slab-dropdown-fixed';
-	public static readonly cssNameCellValue = '.ag-cell-value';
-	public static readonly cssHeaderCellText = '.ag-header-cell-text';
+	public static readonly TAG_NAME_CONTEXTUAL_MENU = 'systelab-context-menu-item';
+	public static readonly CLASS_NAME_DROPDOWN = 'slab-dropdown slab-dropdown-fixed';
+	public static readonly CSS_NAME_CELL_VALUE = '.ag-cell-value';
+	public static readonly CSS_HEADER_CELL_TEXT = '.ag-header-cell-text';
 
 	public static getGridData(gridObject: ElementFinder, col?: string, row?: number): any {
 		if (col === undefined) {
@@ -23,7 +23,7 @@ export class ComponentUtilService {
 		}
 	}
 
-	public static getGridHeaders(gridObject: ElementFinder, col?: number): any {
+	public static getGridHeader(gridObject: ElementFinder, col?: number) {
 		if (col === undefined) {
 			return gridObject.element(by.className('ag-header-container'))
 				.element(by.className('ag-header-row'))
@@ -39,111 +39,24 @@ export class ComponentUtilService {
 	public static getContextMenu(gridObject: ElementFinder, col?: string, row?: number, subIndex?: number) {
 		if (subIndex === undefined) {
 			return this.getGridData(gridObject, col, row)
-				.all(by.tagName(ComponentUtilService.tagNameContextualMenu));
+				.all(by.tagName(ComponentUtilService.TAG_NAME_CONTEXTUAL_MENU));
 		} else {
 			return this.getGridData(gridObject, col, row)
-				.all(by.tagName(ComponentUtilService.tagNameContextualMenu))
+				.all(by.tagName(ComponentUtilService.TAG_NAME_CONTEXTUAL_MENU))
 				.get(subIndex);
 		}
 	}
 
 	/*static getContextualMenuByRow(mainWindow: ElementFinder) {
-	 return mainWindow.all(by.className(ComponentUtilService.classNameDropDown));
+	 return mainWindow.all(by.className(ComponentUtilService.CLASS_NAME_DROPDOWN));
 	 }
 
 	 static getlistProfileHeaders(windowGrid: ElementFinder) {
-	 return windowGrid.all(by.css(ComponentUtilService.cssHeaderCellText));
+	 return windowGrid.all(by.css(ComponentUtilService.CSS_HEADER_CELL_TEXT));
 	 }
 
 	 static getlistCells(windowGrid: ElementFinder) {
-	 return windowGrid.all(by.css(ComponentUtilService.cssNameCellValue));
+	 return windowGrid.all(by.css(ComponentUtilService.CSS_NAME_CELL_VALUE));
 	 }*/
 
-	public static isSorted(arr, arr2?) {
-		// const limit = arr.length - 1;
-		return arr.every(
-			function(_, i) {
-				let bReturn = true;
-
-				if (i > 0) {
-					if (arr2 === undefined) {
-						bReturn = arr[i - 1] <= arr[i];
-					} else {
-						if (arr[i - 1] === arr[i]) {
-							bReturn = arr2[i - 1] <= arr2[i];
-						} else {
-							bReturn = arr[i - 1] < arr[i];
-						}
-					}
-				}
-				return bReturn;
-			}
-		);
-	}
-
-	public static getCurrentTime() {
-		let strReturn = '';
-
-		const today = new Date();
-		const hh = today.getHours();
-		const mm = today.getMinutes();
-		const ss = today.getSeconds();
-
-		if (hh < 10) {
-			strReturn = '0';
-		}
-		strReturn += hh + ':';
-
-		if (mm < 10) {
-			strReturn += '0';
-		}
-		strReturn += mm + ':';
-
-		if (ss < 10) {
-			strReturn += '0';
-		}
-		strReturn += ss;
-
-		return strReturn;
-	}
-
-	public static getCurrentDate() {
-		let strReturn = '';
-
-		const today = new Date();
-		const dd = today.getDate();
-		const mm = today.getMonth() + 1; // January is 0!
-		const yyyy = today.getFullYear();
-
-		if (dd < 10) {
-			strReturn = '0';
-		}
-
-		strReturn += dd + '/';
-		if (mm < 10) {
-			strReturn += '0';
-		}
-
-		strReturn += mm + '/' + yyyy;
-		return strReturn;
-	}
-
-	public static isClickable(el: ElementFinder): promise.Promise<boolean> {
-		return new promise.Promise(resolve => {
-			// let interval = setInterval(() => {
-			// console.log("flag 2");
-			el.click()
-				.then(() => {
-					// clearInterval(interval);
-					setTimeout(() => {
-						// console.log("flag 3");
-						resolve(true);
-					}, 100);
-				}, () => {
-					// console.log("flag 4");
-					resolve(false);
-				});
-			// }, 50);
-		});
-	}
 }
