+++ e2e/common/utilities/test-toolkit.ts
@@ -1,6 +1,7 @@
-import { browser, ElementFinder, ExpectedConditions as EC } from 'protractor';
+import { browser, ElementFinder, ExpectedConditions as EC, promise } from 'protractor';
 import { BasePage } from './base-page';
 import { TestUtil } from './test-util';
+import { JSConsole } from './js-console';
 
 declare const allure: any;
 
@@ -11,6 +12,8 @@ export interface ButtonState {
 }
 
 export class TestToolkit {
+
+	private static console = new JSConsole();
 
 	public static readonly TIME_OUT_MS_FOR_DIALOG_WINDOW = 30000;
 
@@ -30,8 +33,7 @@ export class TestToolkit {
 		if (buttons) {
 			allure.createStep('The buttons are valid', () => {
 			});
-			newPage.checkButtons(buttons);
-
+			this.checkButtons(newPage, buttons);
 		}
 	}
 
@@ -60,5 +62,129 @@ export class TestToolkit {
 					})()
 				});
 		})();
+	}
+
+	public static checkButtons(page: BasePage, buttons: ButtonState[]) {
+		allure.createStep('Action: Get all the buttons', () => {
+		});
+		// Apart, check if the enable flag is correct
+
+		buttons.forEach((buttonElem) => {
+			page.existButton(buttonElem.name)
+				.then((existButton) => {
+					allure.createStep(`Button ${buttonElem.name} is present`, () => {
+						expect(existButton)
+							.toBe(buttonElem.exist, `Button ${buttonElem.name} should be present`);
+					})()
+				});
+		});
+
+		buttons.filter((b) => b.enable)
+			.forEach((buttonElem) => {
+				page.getButtonByName(buttonElem.name)
+					.isEnabled()
+					.then((enabled) => {
+						allure.createStep(`Button ${buttonElem.name} is enabled`, () => {
+						})()
+						expect(enabled)
+							.toBe(true);
+
+					});
+			});
+		buttons.filter((b) => !b.enable)
+			.forEach((buttonElem) => {
+				page.getButtonByName(buttonElem.name)
+					.isEnabled()
+					.then((enabled) => {
+						allure.createStep(`Button ${buttonElem.name} is disabled`, () => {
+						})()
+						expect(enabled)
+							.toBe(null);
+
+					});
+			});
+		page.getAllButtons()
+			.count()
+			.then((inCount) => {
+				allure.createStep(`Buttons count should be equal to ${buttons.filter((b) => b.exist).length}`, () => {
+					expect(inCount)
+						.toBe(buttons.filter((b) => b.exist).length, 'Buttons count');
+				})()
+			});
+
+	}
+
+	public static clearConsole() {
+		return this.console.clear();
+	}
+
+	public static hasErrorsInConsole() {
+		return this.console.hasErrors();
+	}
+
+	public static getCurrentTime() {
+		let strReturn = '';
+
+		const today = new Date();
+		const hh = today.getHours();
+		const mm = today.getMinutes();
+		const ss = today.getSeconds();
+
+		if (hh < 10) {
+			strReturn = '0';
+		}
+		strReturn += hh + ':';
+
+		if (mm < 10) {
+			strReturn += '0';
+		}
+		strReturn += mm + ':';
+
+		if (ss < 10) {
+			strReturn += '0';
+		}
+		strReturn += ss;
+
+		return strReturn;
+	}
+
+	public static getCurrentDate() {
+		let strReturn = '';
+
+		const today = new Date();
+		const dd = today.getDate();
+		const mm = today.getMonth() + 1; // January is 0!
+		const yyyy = today.getFullYear();
+
+		if (dd < 10) {
+			strReturn = '0';
+		}
+
+		strReturn += dd + '/';
+		if (mm < 10) {
+			strReturn += '0';
+		}
+
+		strReturn += mm + '/' + yyyy;
+		return strReturn;
+	}
+
+	public static isClickable(el: ElementFinder): promise.Promise<boolean> {
+		return new promise.Promise(resolve => {
+			// let interval = setInterval(() => {
+			// console.log("flag 2");
+			el.click()
+				.then(() => {
+					// clearInterval(interval);
+					setTimeout(() => {
+						// console.log("flag 3");
+						resolve(true);
+					}, 100);
+				}, () => {
+					// console.log("flag 4");
+					resolve(false);
+				});
+			// }, 50);
+		});
 	}
 }
